+++ creep-tenuous-common/src/main/java/com/zer0s2m/creeptenuous/common/data/DataTransferredUserApi.java
@@ -5,6 +5,6 @@ import io.swagger.v3.oas.annotations.media.Schema;
 public record DataTransferredUserApi(
 
         @Schema(description = "Set Setting - Transfer File Objects to Designated User")
-        Long userId
+        String login
 
 ) { }