+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Removing a color category from a custom category without a unique palette key.
 - Removing a custom category from a file object without specifying a category.
 - Access to view all users for regular roles.
+- Assign file transfer to user using login.
 
 ### Fixed